chore: 프로덕션 패키지 구조에 맞게 테스트 코드 패키지 구조 변경
