+++ src/test/java/com/genius/gitget/challenge/instance/controller/InstanceControllerTest.java
@@ -1,0 +1,5 @@
+package com.genius.gitget.challenge.instance.controller;
+
+
+public class InstanceControllerTest {
+}
