--- src/test/java/com/genius/gitget/file/service/FilesServiceTest.java
@@ -1,25 +1,0 @@
-package com.genius.gitget.file.service;
-
-import com.genius.gitget.global.file.service.FilesService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.transaction.annotation.Transactional;
-
-@SpringBootTest
-@Transactional
-class FilesServiceTest {
-//    @Autowired
-//    private FilesService filesService;
-//
-//    @Test
-//    @DisplayName("")
-//    public void (){
-//        //given
-//
-//        //when
-//
-//        //then
-//
-//    }
-
-}
