+++ src/test/java/com/genius/gitget/global/file/service/FilesServiceTest.java
@@ -1,0 +1,23 @@
+package com.genius.gitget.global.file.service;
+
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.transaction.annotation.Transactional;
+
+@SpringBootTest
+@Transactional
+class FilesServiceTest {
+//    @Autowired
+//    private FilesService filesService;
+//
+//    @Test
+//    @DisplayName("")
+//    public void (){
+//        //given
+//
+//        //when
+//
+//        //then
+//
+//    }
+
+}
