+++ src/test/java/com/genius/gitget/global/security/controller/AuthControllerTest.java
@@ -1,0 +1,53 @@
+package com.genius.gitget.global.security.controller;
+
+import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import com.genius.gitget.challenge.user.domain.Role;
+import com.genius.gitget.util.TokenTestUtil;
+import com.genius.gitget.util.WithMockCustomUser;
+import jakarta.servlet.http.Cookie;
+import lombok.extern.slf4j.Slf4j;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.context.WebApplicationContext;
+
+@SpringBootTest
+@Transactional
+@Slf4j
+public class AuthControllerTest {
+    MockMvc mockMvc;
+    @Autowired
+    WebApplicationContext context;
+
+    @Autowired
+    TokenTestUtil tokenTestUtil;
+
+    @BeforeEach
+    public void setup() {
+        mockMvc = MockMvcBuilders
+                .webAppContextSetup(context)
+                .apply(springSecurity())
+                .build();
+    }
+
+    @Test
+    @DisplayName("anotation test")
+    @WithMockCustomUser(role = Role.USER)
+    public void test() throws Exception {
+        //given
+        Cookie cookie = tokenTestUtil.createAccessCookie();
+
+        //when&then
+        mockMvc.perform(get("/api/test")
+                        .cookie(cookie))
+                .andExpect(status().isOk());
+    }
+}
