--- src/test/java/com/genius/gitget/instance/InstanceControllerTest.java
@@ -1,8 +1,0 @@
-package com.genius.gitget.instance;
-
-
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-
-
-public class InstanceControllerTest {
-}
